feature: add :core:data & :core:testing
