--- app/src/main/java/ksnd/hiraganaconverter/di/module/DataStoreRepositoryModule.kt
@@ -1,17 +1,0 @@
-package ksnd.hiraganaconverter.di.module
-
-import dagger.Binds
-import dagger.Module
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-import ksnd.hiraganaconverter.core.domain.repository.DataStoreRepository
-import ksnd.hiraganaconverter.model.repository.DataStoreRepositoryImpl
-import javax.inject.Singleton
-
-@Module
-@InstallIn(SingletonComponent::class)
-abstract class DataStoreRepositoryModule {
-    @Singleton
-    @Binds
-    abstract fun bindDataStoreRepository(impl: DataStoreRepositoryImpl): DataStoreRepository
-}
