--- app/src/main/java/ksnd/hiraganaconverter/model/ConvertApiClient.kt
@@ -1,13 +1,0 @@
-package ksnd.hiraganaconverter.model
-
-import ksnd.hiraganaconverter.core.model.ResponseData
-import okhttp3.RequestBody
-import retrofit2.Response
-import retrofit2.http.Body
-import retrofit2.http.POST
-
-interface ConvertApiClient {
-
-    @POST(" ")
-    suspend fun requestConvert(@Body body: RequestBody): Response<ResponseData>
-}
