+++ core/data/src/main/java/ksnd/hiraganaconverter/data/client/ConvertApiClient.kt
@@ -1,0 +1,13 @@
+package ksnd.hiraganaconverter.data.client
+
+import ksnd.hiraganaconverter.core.model.ResponseData
+import okhttp3.RequestBody
+import retrofit2.Response
+import retrofit2.http.Body
+import retrofit2.http.POST
+
+interface ConvertApiClient {
+
+    @POST(" ")
+    suspend fun requestConvert(@Body body: RequestBody): Response<ResponseData>
+}
