+++ core/data/src/main/java/ksnd/hiraganaconverter/data/di/DataStoreRepositoryModule.kt
@@ -1,0 +1,17 @@
+package ksnd.hiraganaconverter.data.di
+
+import dagger.Binds
+import dagger.Module
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+import ksnd.hiraganaconverter.core.domain.repository.DataStoreRepository
+import ksnd.hiraganaconverter.data.repository.DataStoreRepositoryImpl
+import javax.inject.Singleton
+
+@Module
+@InstallIn(SingletonComponent::class)
+abstract class DataStoreRepositoryModule {
+    @Singleton
+    @Binds
+    abstract fun bindDataStoreRepository(impl: DataStoreRepositoryImpl): DataStoreRepository
+}
