+++ core/data/src/main/java/ksnd/hiraganaconverter/data/mock/MockConvertHistories.kt
@@ -1,0 +1,11 @@
+package ksnd.hiraganaconverter.data.mock
+
+import ksnd.hiraganaconverter.core.model.ConvertHistoryData
+
+class MockConvertHistories {
+    val data: List<ConvertHistoryData> = listOf(
+        ConvertHistoryData(id = 1, before = "漢字", after = "かんじ", time = "2021/01/01 00:00:00"),
+        ConvertHistoryData(id = 2, before = "漢字", after = "カンジ", time = "2021/01/01 00:00:00"),
+        ConvertHistoryData(id = 3, before = "あいうえお", after = "アイウエオ", time = "2021/01/01 00:00:00"),
+    )
+}
