+++ core/data/src/main/java/ksnd/hiraganaconverter/data/mock/MockConverted.kt
@@ -1,0 +1,6 @@
+package ksnd.hiraganaconverter.data.mock
+
+class MockConverted {
+    val hiragana = "もっくでーた"
+    val katakana = "モックデータ"
+}
