+++ core/resource/src/main/java/ksnd/hiraganaconverter/core/resource/TimeFormat.kt
@@ -1,0 +1,5 @@
+package ksnd.hiraganaconverter.core.resource
+
+enum class TimeFormat(val format: String) {
+    YEAR_MONTH_DATE_HOUR_MINUTE(format = "yyyy/MM/dd HH:mm"),
+}
