+++ settings.gradle.kts
@@ -25,3 +25,5 @@ include(":app")
 include(":core:model")
 include(":core:resource")
 include(":core:domain")
+include(":core:data")
+include(":core:testing")
